@@ -24,21 +24,23 @@
         }
     ],
     "require": {
-        "php": "~5.4|~7.0",
+        "php": "~5.5|~7.0",
 
-        "symfony/http-kernel": "~2.7",
-        "symfony/console": "~2.7",
-        "symfony/config": "~2.7",
-        "symfony/event-dispatcher": "~2.7",
-        "symfony/dependency-injection": "~2.7",
+        "symfony/framework-bundle": "~2.8|~3.0",
+        "symfony/http-kernel": "~2.8|~3.0",
+        "symfony/console": "~2.8|~3.0",
+        "symfony/config": "~2.8|~3.0",
+        "symfony/event-dispatcher": "~2.8|~3.0",
+        "symfony/dependency-injection": "~2.8|~3.0",
+        "symfony/doctrine-bridge": "~2.8|~3.0",
 
+        "doctrine/dbal": "~2.5",
         "doctrine/orm": "~2.5",
         "doctrine/common": "~2.5",
+        "doctrine/doctrine-bundle": "~1.6",
 
-        "tenolo/core-bundle": "~1.0"
-    },
-    "require-dev": {
-        "symfony/phpunit-bridge": "~2.8|~3.0"
+        "mmoreram/symfony-bundle-dependencies": "~1.0|~2.0",
+        "jbzoo/utils": "~1.7"
     },
     "repositories": [
         {
@@ -58,7 +60,9 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-release/1.0": "1.0-dev"
+            "dev-release/1.0": "1.0-dev",
+            "dev-release/2.0": "2.0-dev",
+            "dev-release/3.0": "3.0-dev"
         }
     },
     "minimum-stability": "dev",
